Pull out profile functionality into it's own module. There's not much here yet as there's no managing your own profile or account related functionality
