+++ src/store.js
@@ -1,7 +1,17 @@
-import { applyMiddleware, compose, createStore } from "redux";
+import { applyMiddleware, compose, createStore, combineReducers } from "redux";
 import thunk from "redux-thunk";
 
-import rootReducer from "./reducers";
+import user from "modules/auth/reducer";
+import products from "modules/products/reducer";
+import boxes from "modules/boxes/reducer";
+import profile from "modules/profile/reducer";
+
+const rootReducer = combineReducers({
+  products,
+  profile,
+  user,
+  boxes
+});
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
